+++ alicloud/service_alicloud_pvtz.go
@@ -1,6 +1,8 @@
 package alicloud
 
 import (
+	"strconv"
+
 	"github.com/aliyun/alibaba-cloud-sdk-go/services/pvtz"
 )
 
@@ -27,4 +29,43 @@ func (client *AliyunClient) DescribePvtzZoneInfo(zoneId string) (zone pvtz.Descr
 
 	return
 
+}
+
+func (client *AliyunClient) DescribeZoneRecord(recordId int, zoneId string) (record pvtz.Record, err error) {
+	conn := client.pvtzconn
+	request := pvtz.CreateDescribeZoneRecordsRequest()
+	request.ZoneId = zoneId
+
+	invoker := NewInvoker()
+	err = invoker.Run(func() error {
+		resp, err := conn.DescribeZoneRecords(request)
+
+		recordIdStr := strconv.Itoa(recordId)
+
+		if err != nil {
+			if IsExceptedErrors(err, []string{ZoneNotExists}) {
+				return GetNotFoundErrorFromString(GetNotFoundMessage("PrivateZoneRecord", recordIdStr))
+			}
+			return err
+		}
+		if resp == nil {
+			return GetNotFoundErrorFromString(GetNotFoundMessage("PrivateZoneRecord", recordIdStr))
+		}
+
+		var found bool
+		for _, rec := range resp.Records.Record {
+			if rec.RecordId == recordId {
+				record = rec
+				found = true
+			}
+		}
+
+		if found == false {
+			return GetNotFoundErrorFromString(GetNotFoundMessage("PrivateZoneRecord", recordIdStr))
+		}
+
+		return nil
+	})
+
+	return
 }